apply per-microbatch-loader to only GPT
